ENH: refactor psana*Base clases
Idea is to use psanaBase for handling 1 vs 2 differences,
and move all shared vars and funcs into suiteBase
