+++ suite_scripts/TimeScanParallelSlice.py
@@ -7,9 +7,7 @@
 ## may be copied, modified, propagated, or distributed except according to
 ## the terms contained in the LICENSE.txt file.
 ##############################################################################
-from calibrationSuite.psana2Base import PsanaBase
-#from calibrationSuite.psana1Base import PsanaBase
-
+from calibrationSuite.psanaBase import PsanaBase
 import logging
 import os
 import sys
